Rename spy_matchers.js -> spyMatchers.js
- Rename file, imports, test, snap
- spyMatchers.js file follows internal object, but that object does not
  follow capitalization rule.
- https://github.com/facebook/jest/issues/4969#issuecomment-354961645
  "Files that primarily export types, objects or classes should use
  CapitalizedFileNames.js and should mirror what’s inside 1:1.""
